@@ -16,48 +16,51 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<4.0"
-pandas = "^1.3.3"
-numpy = "^1.21.2"
-scikit-learn = "^1.2.1"
-lxml = "^4.6.3"
-statsbombpy = {version = "^1.3.0", optional = true}
-requests-cache = {version = "^0.9.1", optional = true}
-pandera = "^0.13.4"
+python = ">=3.9,<3.12"
+pandas = "^2.1.1"
+numpy = "^1.26.0"
+scikit-learn = "^1.3.1"
+lxml = "^4.9.3"
+pandera = "^0.17.2"
+statsbombpy = {version = "^1.11.0", optional = true}
+xgboost = {version = "^2.0.0", optional = true}
+tables = {version="^3.8.0", optional = true}
+kloppy = {version=">=3.14.0", optional = true}
 
 [tool.poetry.extras]
-statsbomb = ["statsbombpy", "requests-cache"]
+statsbomb = ["statsbombpy"]
+xgboost = ["xgboost"]
+hdf = ["tables"]
+kloppy = ["kloppy"]
 
-[tool.poetry.dev-dependencies]
-black = "^22.6"
+[tool.poetry.group.dev.dependencies]
+black = "^23.9.1"
+flake8 = "^6.1.0"
+mypy = "^1.5.1"
+pytest = "^7.4.2"
 bumpversion = "^0.6.0"
-flake8 = "^4.0.1"
-mypy = "^0.981"
-pylint = "^2.13.9"
-pytest = "^7.1.2"
-pytest-mock = "^3.8.2"
-unify = "^0.5"
-xgboost = "^1.6.1"
-tables = "^3.6.1"
-Sphinx = "^4.3.2"
-furo = "^2022.6.21"
-tqdm = "^4.64.0"
-coverage = { version="^6.4", extras=["toml"] }
-pytest-cov = "^3.0.0"
-flake8-black = "^0.3.3"
-flake8-bugbear = "^22.7.1"
-flake8-annotations = "^2.9.0"
-flake8-docstrings = "^1.6.0"
+coverage = {version = "^7.3.1", extras = ["toml"]}
+codecov = "^2.1.13"
+pre-commit = "^3.4.0"
+flake8-black = "^0.3.6"
+flake8-bugbear = "^23.9.16"
+flake8-annotations = "^3.0.1"
+flake8-docstrings = "^1.7.0"
 darglint = "^1.8.1"
-flake8-isort = "^4.1.1"
-sphinx-autodoc-typehints = "^1.17.1"
-codecov = "^2.1.12"
-flake8-rst-docstrings = "^0.2.6"
-pre-commit = "^2.19.0"
+flake8-isort = "^6.1.0"
+flake8-rst-docstrings = "^0.3.0"
+pep8-naming = "^0.13.3"
+pyupgrade = "^3.13.0"
+pytest-mock = "^3.11.1"
+pytest-cov = "^4.1.0"
+
+
+[tool.poetry.group.docs.dependencies]
+sphinx = "^7.2.6"
+furo = "^2023.9.10"
 sphinx-autobuild = "^2021.3.14"
-pep8-naming = "^0.13.0"
-Pygments = "^2.12.0"
-pyupgrade = "^2.34.0"
+pygments = "^2.16.1"
+sphinx-autodoc-typehints = "^1.24.0"
 
 [tool.pytest.ini_options]
 filterwarnings = "ignore::DeprecationWarning:tables.*:"
