@@ -1,5 +1,5 @@
 """Implements the feature tranformers of the VAEP framework."""
-from typing import Any, Callable, List, Union
+from typing import Any, Callable, Union
 
 import numpy as np
 import pandas as pd
@@ -42,12 +42,12 @@ __all__ = [
 ]
 
 Actions = Union[DataFrame[SPADLSchema], DataFrame[AtomicSPADLSchema]]
-GameStates = List[Actions]
+GameStates = list[Actions]
 Features = DataFrame[Any]
 FeatureTransfomer = Callable[[GameStates], Features]
 
 
-def feature_column_names(fs: List[FeatureTransfomer], nb_prev_actions: int = 3) -> List[str]:
+def feature_column_names(fs: list[FeatureTransfomer], nb_prev_actions: int = 3) -> list[str]:
     """Return the names of the features generated by a list of transformers.
 
     Parameters
@@ -131,7 +131,7 @@ def actiontype_onehot(actions: Actions) -> Features:
     """
     X = {}
     for type_id, type_name in enumerate(atomicspadl.actiontypes):
-        col = 'type_' + type_name
+        col = 'actiontype_' + type_name
         X[col] = actions['type_id'] == type_id
     return pd.DataFrame(X, index=actions.index)
 
