@@ -1,6 +1,6 @@
 """JSON parser for Opta F24 feeds."""
 from datetime import datetime
-from typing import Any, Dict, Tuple
+from typing import Any
 
 from ...base import MissingDataError
 from .base import OptaJSONParser, _get_end_x, _get_end_y, assertget
@@ -15,13 +15,13 @@ class F24JSONParser(OptaJSONParser):
         Path of the data file.
     """
 
-    def _get_doc(self) -> Dict[str, Any]:
+    def _get_doc(self) -> dict[str, Any]:
         for node in self.root:
             if 'Games' in node['data'].keys():
                 return node
         raise MissingDataError
 
-    def extract_games(self) -> Dict[int, Dict[str, Any]]:
+    def extract_games(self) -> dict[int, dict[str, Any]]:
         """Return a dictionary with all available games.
 
         Returns
@@ -64,7 +64,7 @@ class F24JSONParser(OptaJSONParser):
         }
         return game_dict
 
-    def extract_events(self) -> Dict[Tuple[int, int], Dict[str, Any]]:
+    def extract_events(self) -> dict[tuple[int, int], dict[str, Any]]:
         """Return a dictionary with all available events.
 
         Returns
@@ -92,8 +92,8 @@ class F24JSONParser(OptaJSONParser):
             }
             start_x = float(assertget(attr, 'x'))
             start_y = float(assertget(attr, 'y'))
-            end_x = _get_end_x(qualifiers) or start_x
-            end_y = _get_end_y(qualifiers) or start_y
+            end_x = _get_end_x(qualifiers)
+            end_y = _get_end_y(qualifiers)
 
             event_id = int(assertget(attr, 'id'))
             events[(game_id, event_id)] = dict(
@@ -112,8 +112,8 @@ class F24JSONParser(OptaJSONParser):
                 outcome=bool(int(attr.get('outcome', 1))),
                 start_x=start_x,
                 start_y=start_y,
-                end_x=end_x,
-                end_y=end_y,
+                end_x=end_x if end_x is not None else start_x,
+                end_y=end_y if end_y is not None else start_y,
                 qualifiers=qualifiers,
                 # Optional fields
                 assist=bool(int(attr.get('assist', 0))),
