@@ -1,23 +1,23 @@
 """Implements the feature tranformers of the VAEP framework."""
 from functools import wraps
-from typing import Any, Callable, List, Union, no_type_check
+from typing import Any, Callable, Union, no_type_check
 
 import numpy as np  # type: ignore
 import pandas as pd  # type: ignore
 from pandera.typing import DataFrame
 
-import socceraction.spadl.config as spadlconfig
+import socceraction.spadl.config as spadlcfg
 from socceraction.atomic.spadl import AtomicSPADLSchema
 from socceraction.spadl.schema import SPADLSchema
 
 SPADLActions = DataFrame[SPADLSchema]
 Actions = Union[DataFrame[SPADLSchema], DataFrame[AtomicSPADLSchema]]
-GameStates = List[Actions]
+GameStates = list[Actions]
 Features = DataFrame[Any]
 FeatureTransfomer = Callable[[GameStates], Features]
 
 
-def feature_column_names(fs: List[FeatureTransfomer], nb_prev_actions: int = 3) -> List[str]:
+def feature_column_names(fs: list[FeatureTransfomer], nb_prev_actions: int = 3) -> list[str]:
     """Return the names of the features generated by a list of transformers.
 
     Parameters
@@ -117,9 +117,9 @@ def play_left_to_right(gamestates: GameStates, home_team_id: int) -> GameStates:
     away_idx = a0.team_id != home_team_id
     for actions in gamestates:
         for col in ['start_x', 'end_x']:
-            actions.loc[away_idx, col] = spadlconfig.field_length - actions[away_idx][col].values
+            actions.loc[away_idx, col] = spadlcfg.field_length - actions[away_idx][col].values
         for col in ['start_y', 'end_y']:
-            actions.loc[away_idx, col] = spadlconfig.field_width - actions[away_idx][col].values
+            actions.loc[away_idx, col] = spadlcfg.field_width - actions[away_idx][col].values
     return gamestates
 
 
@@ -139,7 +139,7 @@ def simple(actionfn) -> FeatureTransfomer:
     """
 
     @wraps(actionfn)
-    def _wrapper(gamestates: List[Actions]) -> pd.DataFrame:
+    def _wrapper(gamestates: list[Actions]) -> pd.DataFrame:
         if not isinstance(gamestates, (list,)):
             gamestates = [gamestates]
         X = []
@@ -169,7 +169,13 @@ def actiontype(actions: Actions) -> Features:
     Features
         The 'type_id' of each action.
     """
-    return actions[['type_id']]
+    X = pd.DataFrame(index=actions.index)
+    X["actiontype"] = pd.Categorical(
+        actions["type_id"].replace(spadlcfg.actiontypes_df().type_name.to_dict()),
+        categories=spadlcfg.actiontypes,
+        ordered=False,
+    )
+    return X
 
 
 @simple
@@ -187,8 +193,8 @@ def actiontype_onehot(actions: SPADLActions) -> Features:
         A one-hot encoding of each action's type.
     """
     X = {}
-    for type_id, type_name in enumerate(spadlconfig.actiontypes):
-        col = 'type_' + type_name
+    for type_id, type_name in enumerate(spadlcfg.actiontypes):
+        col = 'actiontype_' + type_name
         X[col] = actions['type_id'] == type_id
     return pd.DataFrame(X, index=actions.index)
 
@@ -207,7 +213,13 @@ def result(actions: SPADLActions) -> Features:
     Features
         The 'result_id' of each action.
     """
-    return actions[['result_id']]
+    X = pd.DataFrame(index=actions.index)
+    X["result"] = pd.Categorical(
+        actions["result_id"].replace(spadlcfg.results_df().result_name.to_dict()),
+        categories=spadlcfg.actiontypes,
+        ordered=False,
+    )
+    return X
 
 
 @simple
@@ -225,7 +237,7 @@ def result_onehot(actions: SPADLActions) -> Features:
         The one-hot encoding of each action's result.
     """
     X = {}
-    for result_id, result_name in enumerate(spadlconfig.results):
+    for result_id, result_name in enumerate(spadlcfg.results):
         col = 'result_' + result_name
         X[col] = actions['result_id'] == result_id
     return pd.DataFrame(X, index=actions.index)
@@ -275,10 +287,18 @@ def bodypart(actions: Actions) -> Features:
     bodypart_detailed :
         An alternative version that splits between the left and right foot.
     """
-    foot_id = spadlconfig.bodyparts.index("foot")
-    left_foot_id = spadlconfig.bodyparts.index("foot_left")
-    right_foot_id = spadlconfig.bodyparts.index("foot_right")
-    return actions[['bodypart_id']].replace([left_foot_id, right_foot_id], foot_id)
+    X = pd.DataFrame(index=actions.index)
+    foot_id = spadlcfg.bodyparts.index("foot")
+    left_foot_id = spadlcfg.bodyparts.index("foot_left")
+    right_foot_id = spadlcfg.bodyparts.index("foot_right")
+    X["bodypart"] = pd.Categorical(
+        actions["bodypart_id"]
+        .replace([left_foot_id, right_foot_id], foot_id)
+        .replace(spadlcfg.bodyparts_df().bodypart_name.to_dict()),
+        categories=["foot", "head", "other", "head/other"],
+        ordered=False,
+    )
+    return X
 
 
 @simple
@@ -303,7 +323,13 @@ def bodypart_detailed(actions: Actions) -> Features:
     bodypart :
         An alternative version that does not split between the left and right foot.
     """
-    return actions[['bodypart_id']]
+    X = pd.DataFrame(index=actions.index)
+    X["bodypart"] = pd.Categorical(
+        actions["bodypart_id"].replace(spadlcfg.bodyparts_df().bodypart_name.to_dict()),
+        categories=spadlcfg.bodyparts,
+        ordered=False,
+    )
+    return X
 
 
 @simple
@@ -328,19 +354,19 @@ def bodypart_onehot(actions: Actions) -> Features:
         An alternative version that splits between the left and right foot.
     """
     X = {}
-    for bodypart_id, bodypart_name in enumerate(spadlconfig.bodyparts):
+    for bodypart_id, bodypart_name in enumerate(spadlcfg.bodyparts):
         if bodypart_name in ('foot_left', 'foot_right'):
             continue
         col = 'bodypart_' + bodypart_name
         if bodypart_name == 'foot':
-            foot_id = spadlconfig.bodyparts.index("foot")
-            left_foot_id = spadlconfig.bodyparts.index("foot_left")
-            right_foot_id = spadlconfig.bodyparts.index("foot_right")
+            foot_id = spadlcfg.bodyparts.index("foot")
+            left_foot_id = spadlcfg.bodyparts.index("foot_left")
+            right_foot_id = spadlcfg.bodyparts.index("foot_right")
             X[col] = actions['bodypart_id'].isin([foot_id, left_foot_id, right_foot_id])
         elif bodypart_name == 'head/other':
-            head_id = spadlconfig.bodyparts.index("head")
-            other_id = spadlconfig.bodyparts.index("other")
-            head_other_id = spadlconfig.bodyparts.index("head/other")
+            head_id = spadlcfg.bodyparts.index("head")
+            other_id = spadlcfg.bodyparts.index("other")
+            head_other_id = spadlcfg.bodyparts.index("head/other")
             X[col] = actions['bodypart_id'].isin([head_id, other_id, head_other_id])
         else:
             X[col] = actions['bodypart_id'] == bodypart_id
@@ -370,17 +396,17 @@ def bodypart_detailed_onehot(actions: Actions) -> Features:
         An alternative version that does not split between the left and right foot.
     """
     X = {}
-    for bodypart_id, bodypart_name in enumerate(spadlconfig.bodyparts):
+    for bodypart_id, bodypart_name in enumerate(spadlcfg.bodyparts):
         col = 'bodypart_' + bodypart_name
         if bodypart_name == 'foot':
-            foot_id = spadlconfig.bodyparts.index("foot")
-            left_foot_id = spadlconfig.bodyparts.index("foot_left")
-            right_foot_id = spadlconfig.bodyparts.index("foot_right")
+            foot_id = spadlcfg.bodyparts.index("foot")
+            left_foot_id = spadlcfg.bodyparts.index("foot_left")
+            right_foot_id = spadlcfg.bodyparts.index("foot_right")
             X[col] = actions['bodypart_id'].isin([foot_id, left_foot_id, right_foot_id])
         elif bodypart_name == 'head/other':
-            head_id = spadlconfig.bodyparts.index("head")
-            other_id = spadlconfig.bodyparts.index("other")
-            head_other_id = spadlconfig.bodyparts.index("head/other")
+            head_id = spadlcfg.bodyparts.index("head")
+            other_id = spadlcfg.bodyparts.index("other")
+            head_other_id = spadlcfg.bodyparts.index("head/other")
             X[col] = actions['bodypart_id'].isin([head_id, other_id, head_other_id])
         else:
             X[col] = actions['bodypart_id'] == bodypart_id
@@ -450,8 +476,8 @@ def endlocation(actions: SPADLActions) -> Features:
     return actions[['end_x', 'end_y']]
 
 
-_goal_x: float = spadlconfig.field_length
-_goal_y: float = spadlconfig.field_width / 2
+_goal_x: float = spadlcfg.field_length
+_goal_y: float = spadlcfg.field_width / 2
 
 
 @simple
@@ -622,10 +648,10 @@ def goalscore(gamestates: GameStates) -> Features:
     actions = gamestates[0]
     teamA = actions['team_id'].values[0]
     goals = actions['type_name'].str.contains('shot') & (
-        actions['result_id'] == spadlconfig.results.index('success')
+        actions['result_id'] == spadlcfg.results.index('success')
     )
     owngoals = actions['type_name'].str.contains('shot') & (
-        actions['result_id'] == spadlconfig.results.index('owngoal')
+        actions['result_id'] == spadlcfg.results.index('owngoal')
     )
     teamisA = actions['team_id'] == teamA
     teamisB = ~teamisA
