@@ -19,7 +19,7 @@ except ImportError:
 
 
 package = "socceraction"
-python_versions = ["3.11", "3.10", "3.9", "3.8"]
+python_versions = ["3.11", "3.10", "3.9"]
 nox.needs_version = ">= 2021.6.6"
 nox.options.sessions = (
     "pre-commit",
@@ -121,7 +121,7 @@ def mypy(session: Session) -> None:
 @session(python=python_versions)
 def tests(session: Session) -> None:
     """Run the test suite."""
-    session.install(".[statsbomb]")
+    session.install(".[statsbomb,kloppy]")
     session.install("coverage[toml]", "pytest", "pygments", "pytest-mock")
     try:
         session.run(
@@ -149,7 +149,7 @@ def coverage(session: Session) -> None:
 def docs_build(session: Session) -> None:
     """Build the documentation."""
     args = session.posargs or ["docs", "docs/_build"]
-    session.install(".")
+    session.install(".[kloppy]")
     session.install("sphinx", "furo", "sphinx-autodoc-typehints")
 
     build_dir = Path("docs", "_build")
@@ -163,7 +163,7 @@ def docs_build(session: Session) -> None:
 def docs(session: Session) -> None:
     """Build and serve the documentation with live reloading on file changes."""
     args = session.posargs or ["--host=0.0.0.0", "docs", "docs/_build"]
-    session.install(".")
+    session.install(".[kloppy]")
     session.install("sphinx", "sphinx-autobuild", "furo", "sphinx-autodoc-typehints")
 
     build_dir = Path("docs", "_build")
