@@ -4,6 +4,7 @@ __all__ = [
     'opta',
     'statsbomb',
     'wyscout',
+    'kloppy',
     'config',
     'SPADLSchema',
     'bodyparts_df',
@@ -13,7 +14,7 @@ __all__ = [
     'play_left_to_right',
 ]
 
-from . import config, opta, statsbomb, wyscout
+from . import config, kloppy, opta, statsbomb, wyscout
 from .config import actiontypes_df, bodyparts_df, results_df
 from .schema import SPADLSchema
 from .utils import add_names, play_left_to_right
